Refactor code

+++ app/Http/Controllers/BookingController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Events\BookingWasCreatedEvent;
 use App\Http\Requests\StoreBookingRequest;
+use App\Http\Services\BookingService;
 use App\Models\Booking;
 use App\Models\Customer;
+use App\Models\Room;
 use App\Models\RoomType;
 use Carbon\Carbon;
 use Gloudemans\Shoppingcart\Facades\Cart;
@@ -14,6 +16,13 @@ use Illuminate\Support\Facades\Auth;
 
 class BookingController extends Controller
 {
+    /**
+     * @param BookingService $bookingService
+     */
+    public function __construct(private readonly BookingService $bookingService)
+    {
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -30,6 +39,17 @@ class BookingController extends Controller
      */
     public function create(Request $request, RoomType $roomType)
     {
+        // Kiểm tra xem còn đủ phòng trống hay không
+        $checkin = Carbon::createFromFormat('d/m/Y', $request->checkin);
+        $checkout = Carbon::createFromFormat('d/m/Y', $request->checkout);
+        $rooms = $this->bookingService->getRoomsAvailable($roomType->id, $checkin, $checkout);
+        if ($rooms->count() < (int) $request->room_total) {
+            alert()->error('Không còn đủ phòng ' . $roomType->name . ' trong khoảng thời gian ' . $request->checkin . ' - ' . $request->checkout)
+                ->showConfirmButton('OK', '#FF7B54');
+
+            return redirect()->back();
+        }
+
         Cart::destroy();
         Cart::add([
             'id' => $roomType->id,
@@ -44,6 +64,7 @@ class BookingController extends Controller
                 'adults' => $request->adults,
                 'children' => $request->children,
                 'images' => $roomType->images,
+                'day_total' => $checkin->diffInDays($checkout),
             ],
         ]);
 
@@ -81,6 +102,17 @@ class BookingController extends Controller
         }
 
         $cart = optional(Cart::content())->first();
+        $checkin = Carbon::createFromFormat('d/m/Y', $cart->options->checkin);
+        $checkout = Carbon::createFromFormat('d/m/Y', $cart->options->checkout);
+
+        $rooms = $this->bookingService->getRoomsAvailable($cart->id, $checkin, $checkout);
+        if ($rooms->isEmpty()) {
+            alert()->error('Không còn phòng trống.')
+                ->showConfirmButton('OK', '#FF7B54')->autoClose(5000);
+
+            return redirect()->route('home');
+        }
+
         $booking = Booking::create([
             'customer_id' => $customer->id,
             'checkin' => Carbon::createFromFormat('d/m/Y', $cart->options->checkin),
@@ -89,15 +121,13 @@ class BookingController extends Controller
             'room_total' => $cart->qty,
             'adults' => $cart->options->adults,
             'children' => $cart->options->children,
+            'money_total' => Cart::total() * $cart->options->day_total,
         ]);
-
-        // Tim phong trong
-        $roomIds = [1];
-        $booking->rooms()->attach($roomIds);
+        $booking->rooms()->attach($rooms->first()->id);
 
         BookingWasCreatedEvent::dispatch($booking);
 
-        alert()->success(__('messages.booking_succesfully'))
+        alert()->success(__('messages.booking_successfully'))
             ->showConfirmButton('OK')->autoClose(5000);
 
         return redirect()->route('home');
@@ -144,5 +174,12 @@ class BookingController extends Controller
             ->latest()->get();
 
         return view('customer.my_booking', compact('bookings'));
+    }
+
+    public function getBookingDetail(Booking $booking)
+    {
+        $booking->load('customer');
+
+        return view('customer.booking_detail', compact('booking'));
     }
 }
